requirements.txt
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pymongo==4.10.1
python-dotenv==1.0.1
Werkzeug==3.0.4
gunicorn==20.1.0
flask-bcrypt==1.0.0
flask-jwt-extended==4.4.4

